smart-text-summerizer/requirements.txt
openai
anthropic
google-genai
python-dotenv
jinja2
loguru